Fixed backwards logic
If the 'noform' query parameter does not exist, then FormFile must be
used because the upload will be a multipart form.  When 'noform' is set,
the uploaded file is plain text in the body of the request and can be
simply copied out to the local file.  Prior to this commit, if 'noform'
was absent from the query, then the multipart form was copied (in
encoded form) to the file.